mv buildUri() to specific handler
