+++ src/sparnatural/components/widgets/AbstractHandler.ts
@@ -60,40 +60,4 @@ export default abstract class AbstractHandler {
       return false;
     }
   
-    buildHttpRequest( ) {
-
-    var headers = new Headers();
-    headers.append(
-        "Accept",
-        "application/sparql-results+json, application/json, */*;q=0.01"
-    );
-    let requestOptions = {
-        method: "GET",
-        headers: headers,
-        mode: "cors",
-        cache: "default",
-    } as {[key:string]:any};
-
-    const config = getSettings().dataEndpoints.find((val)=>{
-    if(val?.endpoint === this.sparqlEndpointUrl)
-    return val
-    })
-    
-    if(config) {
-        // the sparqlEndpointUrl equals an endpoint in Settings.dataEndpoints
-        for (const [key, value] of Object.entries(config)) {
-            if(key === 'endpoint') {
-                if (typeof value !== 'string') {
-                    console.error('endpoint in dataEndpoints must be a string!')
-                }
-                this.sparqlEndpointUrl = value as string
-            } else {
-                requestOptions[key] = value
-            }
-            
-        }
-    } 
-
-    return requestOptions
-    }
   }
